feat: implement cascading deletion utility for child resources
